+++ common/util.h
@@ -14,8 +14,8 @@ void trim_str(std::string &s);
 // replacements for boost::lexical_cast<string>
 // Yes, we could use a template here, but that would generate separate functions
 // for char, unsigned char, short, unsigned short, etc. etc. and all we want are
-// unsigned short and float.
-std::string lex_cast(const unsigned short n);
+// (signed) int and float.
+std::string lex_cast(const int n);
 std::string lex_cast_fl(const float f);
 
 #endif
